==> scripts/analysis/displacement/3 - Analyze accounts.ipynb <==
# Analyze accounts

In this step we look for interesting facts about the accounts.

import json
import os
import tarfile

import matplotlib.pyplot as plt
import pandas as pd
from pandas.plotting import register_matplotlib_converters


register_matplotlib_converters()

%matplotlib inline

plt.style.use("seaborn")

## Load the data

data_directory = os.path.join("..", "..", "..", "data")

results_file_path = os.path.join(data_directory, "displacement_results.json")

if not os.path.exists(results_file_path):
    with tarfile.open(os.path.join(data_directory, "displacement_results.tar.xz"), "r:xz", encoding="utf-8") as compressed_file:
        compressed_file.extract("displacement_results.json", data_directory)

with open(results_file_path, "r", encoding="utf-8") as results_file:
    results = []
    line = results_file.readline().strip()
    while line != "":
        results.append(json.loads(line))
        line = results_file.readline().strip()

with open(os.path.join(data_directory, "displacement_bot_hash_to_cluster_id.json"), "r") as json_file:
    bot_hash_to_cluster_id = json.load(json_file)

with open(os.path.join(data_directory, "displacement_attacker_hash_to_cluster_id.json"), "r") as json_file:
    attacker_hash_to_cluster_id = json.load(json_file)

## Collect into data frames

Wei_to_KWei = 1e3
Wei_to_MWei = 1e6
Wei_to_GWei = 1e9
Wei_to_Ether = 1e18

attackers = []
bots = []
attack_blocks = []
attack_gas_prices = []
victim_senders = []
victim_receivers = []
victim_blocks = []
victim_gas_prices = []
costs = []
profits = []

for result in results:
    same_input = result["attacker_transaction"]["input"] == result["victim_transaction"]["input"]

    # the four entities should be different
    assert len(set([
        result["attacker_transaction"]["from"],
        result["attacker_transaction"]["to"],
        result["victim_transaction"]["from"],
        result["victim_transaction"]["to"]
    ])) == 4

    # the attacker should attack on an earlier block
    # or in the same block but with a lower index
    assert result["attacker_transaction"]["blockNumber"] < result["victim_transaction"]["blockNumber"]\
        or (result["attacker_transaction"]["blockNumber"] == result["victim_transaction"]["blockNumber"]\
        or result["attacker_transaction"]["transactionIndex"] < result["victim_transaction"]["transactionIndex"])

    # the attacker should not pay less for the gas
    assert result["attacker_transaction"]["gasPrice"] >= result["victim_transaction"]["gasPrice"]

    attackers.append(result["attacker_transaction"]["from"])
    bots.append(result["attacker_transaction"]["to"])
    attack_blocks.append(result["attacker_transaction"]["blockNumber"])
    attack_gas_prices.append(result["attacker_transaction"]["gasPrice"] / Wei_to_GWei)

    victim_senders.append(result["victim_transaction"]["from"])
    victim_receivers.append(result["victim_transaction"]["to"])
    victim_blocks.append(result["victim_transaction"]["blockNumber"])
    victim_gas_prices.append(result["victim_transaction"]["gasPrice"] / Wei_to_GWei)

    costs.append(result["cost_usd"])
    profits.append(result["profit_usd"])

# print some statistics
print("# attacks:", len(results))

df_attacks = pd.DataFrame({
    "Attacker": attackers,
    "Bot": bots,
    "Attack Block": attack_blocks,
    "Attack Gas Price": attack_gas_prices,
    "Victim Sender": victim_senders,
    "Victim Receiver": victim_receivers,
    "Victim Block": victim_blocks,
    "Victim Gas Price": victim_gas_prices,
    "Cost": costs,
    "Profit": profits,
    "Bot Cluster ID": [bot_hash_to_cluster_id.get(bot) for bot in bots]  # will give None if bot clusters are not updated
})


df_attacks["Gas Price Delta"] = df_attacks["Attack Gas Price"] - df_attacks["Victim Gas Price"]
df_attacks["Block Delta"] = df_attacks["Victim Block"] - df_attacks["Attack Block"]

df_attacks.to_csv(os.path.join(data_directory, "displacement_attacks.csv"), index=False)

df_attacks

df_bots = df_attacks.groupby("Bot").agg({
        "Bot": "count",
        "Attacker": lambda attacker: attacker.nunique(),
        "Attack Block": "min",
        "Cost": "sum",
        "Profit": "sum",
    })\
    .rename(columns={
        "Bot": "Attacks",
        "Attacker": "Attackers",
        "Attack Block": "First Block",
    })\
    .sort_values(by="Attacks", ascending=False)\
    .reset_index()

df_bots["Cluster ID"] = df_bots["Bot"].map(lambda bot: bot_hash_to_cluster_id.get(bot))  # will give None if bot clusters are not updated

df_bots

df_bots.to_csv(os.path.join(data_directory, "displacement_bots.csv"), index=False)

## Save latex tables for the publication

df_attacks[[
        "Cost",
        "Profit",
        "Gas Price Delta",
        "Block Delta",
    ]].describe().style.format("{:.2f}")

latex = df_attacks[[
        "Cost",
        "Profit",
        "Gas Price Delta",
        "Block Delta",
    ]].describe().to_latex(index=True, formatters={
        "Cost": "{:,.2f}".format,
        "Profit": "{:,.2f}".format,
        "Gas Price Delta": "{:,.2f}".format,
        "Block Delta": "{:,.2f}".format,
    })

with open(os.path.join(data_directory, "displacement_attack_stats.tex"), "w") as latex_file:
    latex_file.write(latex)

## Other stats

### Attack variable correlations

df_attacks[[
        "Cost",
        "Profit",
        "Gas Price Delta",
        "Block Delta",
    ]].corr().style.format("{:.2f}").background_gradient(cmap='coolwarm')
